Rework hub and port enumeration
When a new hub instance is created, the user should be able to enumerate
all attached ports. This is now implemented in the Hub() constructor with
the optional argument `enumerate`. The argument defaults to False.

In order to avoid code deduplication, the port enumeration in discover_hubs()
has been modified to use the new functionality with Hub(path, enumerate=True).

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>