fix: make port handling more robust
The old approach of port discovery and status handling relies on a
certain format of uhubctl output. It turned out that this is kinda
fragile, because it does not always contain the keyword 'off' for a
disabled port (eg. when indicator status is present). Fix this by
removing the status from discovery and adding 'indicator' to status
handling.

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>